Source: https://github.com/msupply-foundation/open-msupply/commit/9ea2c296bdebf77bf83502d65dfc0334cd8c0956

+++ server/graphql/requisition_line/src/mutations/request_requisition_line/update.rs
@@ -18,7 +18,7 @@ use service::{
 #[graphql(name = "UpdateRequestRequisitionLineInput")]
 pub struct UpdateInput {
     pub id: String,
-    pub requested_quantity: Option<u32>,
+    pub requested_quantity: Option<f64>,
     pub comment: Option<String>,
 }
 
@@ -327,7 +327,7 @@ mod test {
                 input,
                 ServiceInput {
                     id: "update line id input".to_string(),
-                    requested_quantity: Some(1),
+                    requested_quantity: Some(1.0),
                     comment: Some("comment".to_string())
                 }
             );
